ckanext/idm/tests/features/requirement.txt
behave==1.2.6
parse==1.12.0
parse-type==0.4.2
psycopg2==2.8.3
PyHamcrest==1.9.0
python-redis==0.2.1
pysolr==3.8.1
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
